WIP: Reorganizes tests (still need to fix travis.yml)
Moves not-unit tests under integration with new flag. Refines test-unit to real unit tests.
