+++ package.json
@@ -5,10 +5,11 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
+    "test-page": "./node_modules/mocha/bin/_mocha test/integration/page",
     "test-integration": "casperjs test ./test/integration/test.*.js",
     "test-integration-runner": "NODE_ENV=test node casper-runner.js",
-    "test": "npm run lint && npm run test-unit && npm run test-integration-runner",
-    "test-unit": "./node_modules/mocha/bin/_mocha",
+    "test": "npm run lint && npm run test-unit && npm run test-page && npm run test-integration-runner",
+    "test-unit": "./node_modules/mocha/bin/_mocha test/unit",
     "lint": "eslint .",
     "autofix": "eslint --fix ."
   },
