+++ test/mocha.opts
@@ -1,4 +1,2 @@
 --compilers js:babel-core/register
 --recursive
-test/views
-test/unit
