+++ test/unit/views/TopStories/test.index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import assert from 'assert';
 import sinon from 'sinon';
 import { shallow } from 'enzyme';
-import TopStories from '../../../views/TopStories/index.jsx';
+import TopStories from '../../../../views/TopStories/index.jsx';
 
 describe('<TopStories/>', () => {
   let stories = [];
